requirements.txt
  django
  python-dotenv>=0.19.2